# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "eCommerce",
    "description": "Documentation for eCommerce Exam Project"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Get all products that are not deleted in the database",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Create a new product",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateProduct"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Update a product",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateProduct"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Soft delete a product (setting IsDeleted = 1)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/DeleteProduct"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/memberships/": {
      "get": {
        "tags": [
          "Membership"
        ],
        "description": "Get all membership levels",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Membership"
        ],
        "description": "Create new membership",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateMembership"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Membership"
        ],
        "description": "Update one of the current membership levels",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateMembership"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Membership"
        ],
        "description": "Deleting a membership level, only if it do not belong to a user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/DeleteMembership"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/admin/products": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Get admin products page with crud functionality",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/admin/products/{data}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Searching for product name, brand or category using the search bar in the products page",
        "parameters": [
          {
            "name": "data",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/admin/memberships": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Get membership page for admin with crud functionality",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateProduct": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string",
          "example": "Any text"
        },
        "Description": {
          "type": "string",
          "example": "Any text"
        },
        "UnitPrice": {
          "type": "string",
          "example": "999"
        },
        "imgurl": {
          "type": "string",
          "example": "Any text"
        },
        "Quantity": {
          "type": "string",
          "example": "10"
        },
        "BrandId": {
          "type": "string",
          "example": "1"
        },
        "CategoryId": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "Name",
        "Description",
        "UnitPrice",
        "imgurl",
        "Quantity",
        "BrandId",
        "CategoryId"
      ]
    },
    "UpdateProduct": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "17"
        },
        "Name": {
          "type": "string",
          "example": "Updated Product Name"
        },
        "Description": {
          "type": "string",
          "example": "Updated"
        },
        "UnitPrice": {
          "type": "string",
          "example": "649"
        },
        "imgurl": {
          "type": "string",
          "example": "Test"
        },
        "Quantity": {
          "type": "string",
          "example": "10"
        },
        "IsDeleted": {
          "type": "string",
          "example": "0"
        },
        "BrandId": {
          "type": "string",
          "example": "1"
        },
        "CategoryId": {
          "type": "string",
          "example": "5"
        }
      },
      "required": [
        "id",
        "Name",
        "Description",
        "UnitPrice",
        "imgurl",
        "Quantity",
        "IsDeleted",
        "BrandId",
        "CategoryId"
      ]
    },
    "DeleteProduct": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "id"
      ]
    },
...
    "UpdateMembership": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1"
        },
        "newStatus": {
          "type": "string",
          "example": "UpdatedStatusName"
        },
        "newDiscount": {
          "type": "string",
          "example": "40"
        }
      },
      "required": [
        "id",
        "newStatus",
        "newDiscount"
      ]
    },
    "DeleteMembership": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1"
        }
      },
      "required": [
        "id"
      ]
    },
...
  }
}